Move useSarcophagusNegotion out of libp2p folder
